User experience updates. 1.Fully implemented LT & ENG languages. 2. Standartized Thymeleaf design in all pages.

+++ src/main/java/lt/codeacademy/projects/chc/coronahatersclub/service/PostService.java
@@ -4,10 +4,13 @@ import lombok.RequiredArgsConstructor;
 import lt.codeacademy.projects.chc.coronahatersclub.model.Post;
 import lt.codeacademy.projects.chc.coronahatersclub.model.User;
 import lt.codeacademy.projects.chc.coronahatersclub.repository.PostRepository;
-import lt.codeacademy.projects.chc.coronahatersclub.repository.UserRepository;
+import lt.codeacademy.projects.chc.coronahatersclub.validators.PostDeleteValidator;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 
+import javax.persistence.EntityNotFoundException;
 import java.time.LocalDateTime;
 import java.time.ZoneId;
 import java.util.List;
@@ -16,7 +19,7 @@ import java.util.List;
 @RequiredArgsConstructor
 public class PostService {
     private final PostRepository postRepository;
-    private final UserRepository userRepository;
+    private final PostDeleteValidator postDeleteValidator;
 
     public String createNewPost(Authentication authentication, Post post) {
         User user = (User)authentication.getPrincipal();
@@ -32,7 +35,22 @@ public class PostService {
         return postRepository.findAllByUser(user);
     }
 
-    public void deletePost(Long postId) {
-        postRepository.deleteById(postId);
+    public String deletePost(Long postId, Authentication authentication) {
+        User user = (User) authentication.getPrincipal();
+        Post post = postRepository.findById(postId).orElseThrow(()->new EntityNotFoundException("Post with given Id does not exist"));
+        boolean deleteValid = postDeleteValidator.validate(user,post);
+        if(deleteValid) {
+            postRepository.deleteById(postId);
+        }
+        return "redirect:/user/profile/posts";
+    }
+
+    public Page<Post> getAllPostsPageable(Integer page) {
+        return postRepository.findAllOrderByDate(PageRequest.of(page,5));
+    }
+
+    public Page<Post> getThreePosts() {
+        return postRepository.findAllOrderByDate(PageRequest.of(0,3));
+
     }
 }
